+++ lib/flutter_widget_recorder.dart
@@ -1,4 +1,5 @@
 library;
 
+export 'src/recording_options.dart';
 export 'src/widget_recorder_controller.dart';
 export 'src/widget_recorder_wrapper.dart';