+++ test/flutter_widget_recorder_method_channel_test.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_widget_recorder/src/flutter_widget_recorder_method_channel.dart';
+import 'package:flutter_widget_recorder/src/recording_options.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
@@ -8,12 +9,14 @@ void main() {
   MethodChannelFlutterWidgetRecorder platform =
       MethodChannelFlutterWidgetRecorder();
   const MethodChannel channel = MethodChannel('flutter_widget_recorder');
+  MethodCall? lastCall;
 
   setUp(() {
     TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
         .setMockMethodCallHandler(
       channel,
       (MethodCall methodCall) async {
+        lastCall = methodCall;
         switch (methodCall.method) {
           case 'startRecording':
             return true;
@@ -46,6 +49,36 @@ void main() {
       ),
       true,
     );
+    expect(lastCall?.arguments, {
+      'name': 'test',
+      'width': 100,
+      'height': 100,
+      'pixelRatio': 1.0,
+    });
+  });
+
+  test('startRecording forwards optional encoder args', () async {
+    await platform.startRecording(
+      name: 'test',
+      width: 100,
+      height: 100,
+      pixelRatio: 1.0,
+      targetFps: 60,
+      bitrateBps: 4000000,
+      iFrameIntervalSec: 2,
+      bitrateMode: RecorderBitrateMode.vbr,
+    );
+
+    expect(lastCall?.arguments, {
+      'name': 'test',
+      'width': 100,
+      'height': 100,
+      'pixelRatio': 1.0,
+      'targetFps': 60,
+      'bitrateBps': 4000000,
+      'iFrameIntervalSec': 2,
+      'bitrateMode': 'vbr',
+    });
   });
 
   test('stopRecording', () async {