+++ test/flutter_widget_recorder_test.dart
@@ -4,6 +4,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_widget_recorder/src/flutter_widget_recorder.dart';
 import 'package:flutter_widget_recorder/src/flutter_widget_recorder_method_channel.dart';
 import 'package:flutter_widget_recorder/src/flutter_widget_recorder_platform_interface.dart';
+import 'package:flutter_widget_recorder/src/recording_options.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
 class MockFlutterWidgetRecorderPlatform
@@ -24,6 +25,10 @@ class MockFlutterWidgetRecorderPlatform
     required int width,
     required int height,
     required double pixelRatio,
+    int? targetFps,
+    int? bitrateBps,
+    int? iFrameIntervalSec,
+    RecorderBitrateMode? bitrateMode,
   }) async {
     return true;
   }